make state handling the localest
