+++ sessions/react-global-state/demo-start/components/CounterList.js
@@ -1,6 +1,5 @@
 import styled from "styled-components";
 import Counter from "./Counter";
-import { useState } from "react";
 
 const List = styled.ul`
   display: flex;
@@ -11,47 +10,15 @@ const List = styled.ul`
   gap: 1rem;
 `;
 
-const initialAnimalArray = [
-  { id: 1, name: "Cats", count: 0 },
-  { id: 2, name: "Dogs", count: 0 },
-  { id: 3, name: "Sheep", count: 0 },
-  { id: 4, name: "Dragons", count: 0 },
-];
-
 export default function CounterList() {
-  const [animals, setAnimals] = useState(initialAnimalArray);
-
-  function handleAdd(animalID) {
-    setAnimals(
-      animals.map((animal) =>
-        animal.id === animalID ? { ...animal, count: animal.count + 1 } : animal
-      )
-    );
-  }
-
-  function handleSubtract(animalID) {
-    setAnimals(
-      animals.map((animal) =>
-        animal.id === animalID
-          ? { ...animal, count: Math.max(0, animal.count - 1) }
-          : animal
-      )
-    );
-  }
-
   return (
     <>
       <h2>Counters</h2>
       <List>
-        {animals.map((animal) => (
-          <li key={animal.id}>
-            <Counter
-              animal={animal}
-              handleAdd={handleAdd}
-              handleSubtract={handleSubtract}
-            />
-          </li>
-        ))}
+        <Counter animalName={"Cats"} />
+        <Counter animalName={"Dogs"} />
+        <Counter animalName={"Sheep"} />
+        <Counter animalName={"Dragons"} />
       </List>
     </>
   );
